app/accounts/api/urls.py
from django.urls import path
from rest_framework.authtoken.views import obtain_auth_token
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from .views import Register

urlpatterns = [
    path("register/", Register.as_view(), name="register"),
    # path("register/", register, name="register"),
    # path("login/", obtain_auth_token, name="login"),
    path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
