requirements.txt
Django>=4.2.3,<4.3
django-environ>=0.10.0,<0.11
django-filter>=23.2,<23.3
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.2.2,<5.3
drf-spectacular>=0.26.3,<0.27
flake8>=5.0.4,<5.1
psycopg>=3.1.9,<3.2